requirements.txt
fastapi
uvicorn
python-multipart
Pillow
numpy
scikit-learn
matplotlib
seaborn
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
torchvision==0.17.0+cpu
